+++ src/components/ModelIcon/index.tsx
@@ -45,9 +45,7 @@ const ModelIcon = memo<ModelProviderIconProps>(({ model: originModel, size = 12 
   // currently supported models, maybe not in its own provider
   if (model.includes('gpt-3')) return <OpenAI.Avatar size={size} type={'gpt3'} />;
   if (model.includes('gpt-4')) return <OpenAI.Avatar size={size} type={'gpt4'} />;
-  if (model.startsWith('glm') ||
-    model.includes('chatglm'))
-    return <ChatGLM.Avatar size={size} />;
+  if (model.startsWith('glm') || model.includes('chatglm')) return <ChatGLM.Avatar size={size} />;
   if (model.includes('claude')) return <Claude.Avatar size={size} />;
   if (model.includes('titan')) return <Aws.Avatar size={size} />;
   if (model.includes('llama')) return <Meta.Avatar size={size} />;
@@ -56,13 +54,9 @@ const ModelIcon = memo<ModelProviderIconProps>(({ model: originModel, size = 12 
   if (model.includes('gemma')) return <Gemma.Avatar size={size} />;
   if (model.includes('moonshot')) return <Moonshot.Avatar size={size} />;
   if (model.includes('qwen')) return <Tongyi.Avatar background={Tongyi.colorPrimary} size={size} />;
-  if (model.includes('minmax')) return <Minimax.Avatar size={size} />;
-  if (model.includes('mistral') ||
-    model.includes('mixtral'))
-    return <Mistral.Avatar size={size} />;
-  if (model.includes('pplx') ||
-    model.includes('sonar'))
-    return <Perplexity.Avatar size={size} />;
+  if (model.includes('minmax') || model.includes('abab')) return <Minimax.Avatar size={size} />;
+  if (model.includes('mistral') || model.includes('mixtral')) return <Mistral.Avatar size={size} />;
+  if (model.includes('pplx') || model.includes('sonar')) return <Perplexity.Avatar size={size} />;
   if (model.includes('yi-')) return <Yi.Avatar size={size} />;
   if (model.startsWith('openrouter')) return <OpenRouter.Avatar size={size} />; // only for Cinematika and Auto
   if (model.startsWith('openchat')) return <OpenChat.Avatar size={size} />;
@@ -73,13 +67,11 @@ const ModelIcon = memo<ModelProviderIconProps>(({ model: originModel, size = 12 
   if (model.includes('baichuan'))
     return <Baichuan.Avatar background={Baichuan.colorPrimary} size={size} />;
   if (model.includes('rwkv')) return <Rwkv.Avatar size={size} />;
-  if (model.includes('ernie'))
-    return <Wenxin.Avatar size={size} />;
+  if (model.includes('ernie')) return <Wenxin.Avatar size={size} />;
   if (model.includes('spark')) return <Spark.Avatar size={size} />;
   if (model.includes('hunyuan')) return <Hunyuan.Avatar size={size} />;
   // ref https://github.com/fishaudio/Bert-VITS2/blob/master/train_ms.py#L702
-  if (model.startsWith('d_') ||
-    model.startsWith('g_') || model.startsWith('wd_'))
+  if (model.startsWith('d_') || model.startsWith('g_') || model.startsWith('wd_'))
     return <FishAudio.Avatar size={size} />;
   if (model.includes('skylark')) return <ByteDance.Avatar size={size} />;
 
@@ -97,9 +89,7 @@ const ModelIcon = memo<ModelProviderIconProps>(({ model: originModel, size = 12 
   if (model.includes('wizardlm')) return <Azure.Avatar size={size} />;
   if (model.includes('phi3')) return <Azure.Avatar size={size} />;
   if (model.includes('firefly')) return <Adobe.Avatar size={size} />;
-  if (model.includes('jamba') ||
-    model.includes('j2-'))
-    return <Ai21.Avatar size={size} />;
+  if (model.includes('jamba') || model.includes('j2-')) return <Ai21.Avatar size={size} />;
 });
 
 export default ModelIcon;